requirements.txt
# Core agent & LangGraph libraries
langgraph>=0.3.18
langchain-google-genai>=2.0.10
google-genai>=1.9.0
google-adk>=0.0.3
asyncio>=3.4.3
click>=8.1.8
# HTTP & streaming
httpx>=0.28.1
httpx-sse>=0.4.0
sse-starlette>=2.2.1
starlette>=0.46.1
uvicorn==0.34.0
langchain-deepseek==0.1.3
# CLI support
asyncclick>=8.1.8

# Data validation and typing
pydantic>=2.10.6

# Push notification security
jwcrypto>=1.5.6
pyjwt>=2.10.1

# Optional: Streamlit or UI tools (comment out if not needed)
streamlit==1.44.0
uvicorn>=0.34.0

# Optional Dev Tools
# pytest==8.3.5
# pytest-mock==3.14.0
# ruff==0.11.2